1) BUG DESCRIPTION- white margins on the right and bottom of the chat page in chrome.
BUG SOLUTION-
Adding to body overflow: hidden.
Not sure why it helped but it did.

2) BUG DESCRIPTION-
Input was overflowing container to the right on small screens.

BUG SOLUTION-
Set the input width to 50px which made it responsive.
Again not sure why it helped. I guess that the input has some kind of rigid minimum width if not specified other width.